Getting ready for 2D rendering

+++ Epoch/include/Epoch/Graphics/Buffer.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "Epoch/Core.h"
+#include "Epoch/Core/Core.h"
 
 namespace Epoch {
 	
@@ -15,7 +15,7 @@ namespace Epoch {
 		std::string Name;
 		ShaderDataType Type;
 		uint32_t Size;
-		uint32_t Offset;
+		size_t Offset;
 		bool Normalized;
 
 		BufferElement() { }
